most of it working. Still no modal. working now...
