+++ lambda-notes/package.json
@@ -1,17 +1,15 @@
 {
-  "name": "lambda-notes",
+  "name": "lambda_notes",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.0.0",
-    "prop-types": "^15.6.1",
+    "axios": "^0.18.0",
+    "bootstrap": "^4.1.0",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
     "react-redux": "^5.0.7",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.4",
-    "reactstrap": "^5.0.0-beta.3",
+    "reactstrap": "^5.0.0",
     "redux": "^4.0.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0"
