+++ lambda-notes/src/actions/index.js
@@ -1,3 +1,4 @@
+
 export const ADD_NOTE = "ADD_NOTE";
 export const DELETE_NOTE = "DELETE_NOTE";
 export const EDIT_NOTE = "EDIT_NOTE";
